feat(icons): configuration now via shadowing like theme-ui
