+++ examples/full/src/gatsby-theme-mdx-suite-base/icons.js
@@ -1,0 +1,5 @@
+import icons from 'gatsby-theme-mdx-suite-base/src/icons'
+
+import GatsbyMonogram from '../assets/icons/gatsby-monogram.svg'
+
+export default { ...icons, gatsby: { icon: GatsbyMonogram } }
